crew/knowledge_base_crew/pyproject.toml
[project]
name = "knowledge_base_crew"
version = "0.1.0"
description = "knowledge-base-crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.140.0,<1.0.0",
    "opensearch-py>=2.0.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
knowledge_base_crew = "knowledge_base_crew.main:run"
run_crew = "knowledge_base_crew.main:run"
train = "knowledge_base_crew.main:train"
replay = "knowledge_base_crew.main:replay"
test = "knowledge_base_crew.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
